operations: Fix flakiness due to EOF timeout
In our merge-related tests, we're using testhelper.ReceiveEOFWithTimeout
to wait for the EOF error. If we fail to receive it after more than a
second, then the test fails. It's debatable what the real merit of this
timeout is: the code is not timing sensitive, but we just want to assert
that it does indeed succeed. Furthermore, given our recent shift to use
`t.Parallel()` more, we have seen that this timeout may be exceeded on
occasion.

Remove these timeouts and simply assert that we do get an EOF error and
drop the now-unused helper function.