repository: Return errors if replicating config fails
Back when we introduced replication of the gitconfig in c28412f83
(repository: Replicate gitconfig, 2021-05-19) , the GetConfig RPC was
introduced to support this replication in the first place. We thus
couldn't count on servers supporting the RPC call yet and had to swallow
errors returned by GetConfig to not cause replication to fail whenever
talking to an oldish server.

This code has long since been released, and thus we can now remove the
"hack" and instead return error codes from GetConfig to the caller.

Changelog: fixed