repository: Fix wrong error code if replication source does not exist
When replicating a repository, then we need to replicate refs as well as
the gitconfig and gitattributes files. To speed up things a bit, these
synchronization functions are run in parallel. In case replication fails
though because the source repository does not exist, we have observed in
CI that from time to time we get an "Internal" error code while other
times we get the correct "NotFound" error.

The issue is that depending on which synchronization function runs
first, we may get different behaviour. While syncing the gitconfig and
gitattributes files knows to correctly pass up the error code, syncing
refs requires an internal fetch via gitaly-ssh. Given that we cannot
pass error codes from gitaly-ssh to git-fetch(1) to Gitaly, we thus
loose the error code and end up with an "Internal" error.

Ideally, we'd fix this issue by propagating error codes by gitaly-ssh.
But this is likely non-trivial fix given that there is no obvious
channel which we could use for this: git-fetch(1) uses `die()` in case
gitaly-ssh fails, which means we cannot return the exit code. Parsing
output of git-fetch(1) seems ugly, too. And opening a sidechannel feels
excessive for this case.

Instead, we just fix this the pragmatic way by doing a RepositoryExists
call previous to starting the synchronization. The single added call
shouldn't matter in the context of the likely-expensive execution of
git-fetch(1) anyway, and like this we can definitely guarantee
consistency for this case.

Changelog: fixed