Fix TestReplicatorInvalidSourceRepository test
TestReplicatorInvalidSourceRepository test hasn't been working correctly
as it does not assert a record exists for the repository prior to deleting
the record and asserting a record doesn't exist. SetGeneration was updated
earlier not to create records in the 'repositories' table anymore. Since then,
the SetGeneration call in the test hasn't been enough to create the database
record the test expects to delete. This commit fixes the issue by replacing
the 'SetGeneration' call with a 'CreateRepository' call that creates the
database record and asserts the record actually exists prior to testing the
deletion.