repository: Return gRPC error codes if syncing files fails
When writing files with the new extended transactional semantics, then
we wrap returned errors via `fmt.Errorf()`. In case the reader fails
though because of an gRPC error, then we now lose the error code
compared to the old code paths which returned the errors as-is. Our
tests didn't catch this bug given that they overwrote any feature flags
set up in the context when injecting the Gitaly server metadata.

Fix the issue by returning errors as-is to fix this and properly merge
the outgoing metadata in tests to catch any future regressions.