testhelper: Add function to quarantine a test
We are having issues with some tests being flaky and blocking deploys
because the pipeline on master is failing. This annoying, but often it's
not easy to quickly fix the flaky test.

This change allows us to put flaky tests in quarantine. The quarantined
tests are run in a separate job, but that job is marked to allow to
fail. This allows us to make the pipeline green again quickly, but still
have the tests running.