Add a method for getting consistent storages by repository ID
Praefect needs to get consistent storages when routing mutators to
repositories. Currently the only way to do so is by the virtual storage
and relative path which may change during the different queries in the
router. This commit adds a method for getting the consistent storages by
a repository ID so the routing won't fail due to races with other RPCs
such as the rename.