ci: Default image uses default versions
The CI yaml file specifies default versions for tools like Ruby and
Golang. But the default image still uses older versions of these tools.

This change uses the default versions of these tools in the default
image. This removes the need the need to specify the image in many jobs.