list-untracked-repositories: Introduce repository Walker struct
The Walker struct introduced to iterate over the storages of
the gitaly cluster and run abstract operation on top of the
found repositories. The ExecOnRepositories method accepts a
callback function that is called once the batch is ready for
the processing. If the callback function returns an error,
ExecOnRepositories stops iterating and returns an error back
to the caller.