Makefile: Never ruby bundle as deployment
During the bundle step of the gitaly-ruby gems it tries to guess whether
it is part of a gitlab-development-kit. If it isn't it installs gems as
a "deployment", which will install the gems in vendor/bundle instead of
in $GEM_HOME. To detect this, it looks for .gdk-install-root in the
parent directory. This behavior is unexpected to many and also seems to
cause issues for some users.

Added to that, recently a change [1] was made to gitlab-org/gitlab to
install gems in $GEM_HOME when running tests. This speeds up the
installation of Gitaly. This change made it so the gems are never
installed as a "deployment"

This change removes the code to determine whether it's a deployment and
makes the gems install in $GEM_HOME unconditionally.

1. https://gitlab.com/gitlab-org/gitlab/-/commit/77f41fff0363ec0b1bc687e8df02995729e8218e

Issue: https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1315