list-untracked-repositories: New praefect sub-command
Praefect extended with a new sub-command 'list-untracked-repositories'.
It walks over all repositories of all accessible gitaly nodes and
checks if repositories exist in the praefect database. If repository
doesn't exist in the database the information about its location
will be printed on STDOUT in JSON format terminated with a newline
(it is configurable with --delimiter flag). STDERR is used to write
log messages and errors if any.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3792