ref: Add new FindRefsByOID RPC method
Adds the proto definition and generated go and ruby code for a new RPC
FindRefsByOID used to find refs that point to a specific object ID with
some flags: patterns the refs need to match, a limit of how many results
are returned, and a field by which to sort the results.

See https://git-scm.com/docs/git-for-each-ref as a reference for how
these flags are used in git for-each-ref.

Changelog: added