testserver: Precreate socket to fix health check race with Praefect
When running the test-with-praefect target, then we're running the
Praefect executable as a transparent proxy in front of Gitaly servers.
We frequently observe test failures though because we fail to establish
the server's healthiness, where the error message typically indicates
that the server's socket path does not exist. This may have two root
causes: either we're really taking too long to spawn Praefect, or we
just fail to dial Praefect correctly in case the socket really hasn't
been created yet.

Plug one of these two potential races by precreating the Praefect socket
path. If this doesn't fix the issue, then the only thing left we can do
is to bump the three second timeout.