testserver: Increase timeout when waiting for healthiness
Most of the test failures we're observing in our CI pipelines right now
happen due to Praefect not becoming healthy. This isn't much of a
surprise: we need to spawn the Praefect executable, Praefect needs to do
its initial setup, connect to the database... and this all needs to
finish in at most three seconds, which feels quite optimistic if the
machine is highly loaded.

Increase the timeout to 15 seconds. While this may be excessive, I'd
rather be safe and have less CI failures.