remote: Simplify interception of gitaly-ssh invocations
In order to test invocation of gitaly-ssh, we intercept calls to it via
a script which then invokes gitaly-ssh. This script will record command
line arguments and the environment of the invocation such that we can
verify that they match our expectations. Setup of this interception and
extraction of written files is quite complex though given that we
support multiple invocations of it, where each invocation writes a
separate, indexed file. As it turns out, this complexity is completely
unneeded: there is only a single test which makes use of this
interception, and this function will only ever execute gitaly-ssh once.

Simplify interception by only supporting a single invocation. While at
it, this commit also sets the `-e` option in the script such that we
detect any failing commands, which was a gap in our setup previously.