Return replica path from GetConsistentStorages
Praefect will start generating unique relative paths for repositories in
14.6 to prevent stale state conflicting with recreated repositories. To
facilitate this, the 'repositories' table has the 'replica_path' column
to store where the replicas of a repository are stored. While in 14.5
Praefect still doesn't generate the paths, it needs to route the requests
to the paths stored in the database. This allows 14.6 then to start generating
unique paths and be backwards compatible wiht 14.5 as it is already using
the paths from the database.

This commit returns the replica path from GetConsistentStorages. The method
was chosen mostly as its result is cached for accessors, so integrating the
replica path into the cache is effortless. This commit only includes the changes
required to fetch the replica path and cache it. A later commit will update the
routing logic to actually route the requests to the stored path.