Rewrite messages to use replica path as the relative path
This commit rewrites the relative path of repositories in the routed
RPCs to the replica path. Currently the replica path is always the same
as the relative path the clients use to access repositories. Praefect
will soon begin generating unique paths for repositories. Using the
replica paths from the database in 14.5 allows generating unique paths
later in a backwards compatible manner.