testserver: Simplify creation of internal socket dir
When an internal socket directory is set when running the Gitaly server,
then we will create the socket directory if it doesn't exist. The logic
to do so is needlessly complex though, where we try to not create it if
it exists already. The intent may be to not schedule the `t.Cleanup()`
call which removes the directory if it was preexisting already. But the
path shouldn't ever point to anything but a temporary directory anyway,
and cleaning those up after the test is done is not going to cause any
harm even if it was set up by the caller.

Refactor the tests to just always set up the internal socket directory.