testserver: Fix setup of the internal server
When setting up the Gitaly test server, then we need to serve both the
internal socket as well as the external socket. Right now, we do set up
both on the same external server though, which is wrong: first, the
shutdown order is different such that external servers are always shut
down before internal servers. And second, the internal server shouldn't
ever use TLS.

We haven't noticed this until now given that we never rely on shutdown
order and because we don't ever seem to connect to internal sockets when
TLS is set up. But it keeps us from waiting on the internal socket to
become healthy.

Fix the setup to use a separate internal server for the internal socket.