supervisor: Simplify test setup
The supervisor tests are a bit antiquated with regards to their code
style, and it's not easy to see what's going on withouth having deep
knowledge about what the supervisor does.

Refactor the code to be easier to read and add a bunch of comments
explaining what happens. While at it, this also adds another testcase
which explicitly tests `NewConfigFromEnv()`.