supervisor: Synchronize on process startup before connecting
The way the supervisor tests determine whether the process has started
already is by just plainly trying to connect to its socket until it
succeeds. There's a better way though via the event notification
channel, which causes the supervisor to send us an event in case the
executable was spawned. This allows us to tighten the test in case we
got an Up event given that we can now fail hard if connecting to the
service fails. And it verifies that the supervisor indeed knows to send
the event, increasing test coverage.

Convert the code to use the event notification channel. Hopefully, this
change fixes test flakiness in this package. But if not, then we should
now get better diagnostics about why tests failed.