gitpipe: Convert `RevisionFilter()` into an option
The `RevisionFilter()` pipeline step is currently an extra pipeline step
with its own set of channels. This is inefficient given that we have an
extra pair of of channels on which we have to receive data from and send
data on. Given that filtering itself should typically be fast, the added
complexity isn't really worth it at all.

Convert the filter into an option for the `Revlist()` pipeline step such
that it runs before sending it down the pipeline. Given that it's been
unobvious whether "filtering" needs to return `true` or `false` to skip
over an object, the new option now has "skipping" semantics, which
should avoid that confusion. Note that this reverses the filter logic
though.