gitpipe: Allow specifying revisions in `ForEachRef()` step
Add a new option which allows the caller to specify the format to be
used by git-for-each-ref(1). While the actual format must always be in
the format '%(objectname) %(refname)', the caller can play some tricks
to inject additional lines into the format via conditional format items
or may change the refname, if he intends to.

The new option is not yet used.