ref: Refactor FindAllTags to skip `CatfileInfo()` pipeline step
The `FindAllTags()` RPC call has four pipeline steps:

    1. It executes git-for-each-ref(1) to enumerate all refs.

    2. The output of git-for-each-ref(1) is transformed such that for
       each tag, we request object info both of the tag and its peeled
       non-tag object via '^{}'.

    3. git-cat-file(1) is used to see whether any ref is an annotated
       tag or not. If the current object is an annotated tag, then we
       pass along both the annotated tag and its peeled object info.
       Otherwise, we skip sending along the peeled object info.

    4. git-cat-file(1) will then retrieve objects. If it's a tag, it
       knows to also request the second object and parse it such that we
       can return both information about the tag, and the object it
       references.

This is a lot of work to do, and FindAllTags is an RPC that's executed
quite often. In repos with many tags, it's shown to be hugely expensive
with lots of small allocations and bumping back-and-forth between the
Goroutines.

We can get rid of the second and third pipeline steps though by playing
a trick: git-for-each-ref(1) has the ability to conditionally print
things depending on what the current object is. Most importantly, it
allows us to print another line if the object type is a tag. We can thus
combine steps one and two into a single step by always printing the
current object, and printing the current object's OID with "^{}"
appended if the current object is a tag. Furthermore, this allow us to
skip inspecting the object type and filtering in the third step given
that git-for-each-ref(1) has already asserted object types for us.

Convert the code to do this. This both reduces the pipeline steps by
half and gets rid of one of two git-cat-file(1) processes. Together, we
should see a notable performance improvement with this.

Changelog: performance