gitpipe: Allow using iterators interchangeably as input
The pipeline architecture in the gitpipe package is currently quite
strict given that we cannot use e.g. the `Revlist()` iterator as input
to the `CatfileObject()` step. It's thus hard to mix-and-match pipeline
steps.

Lift this limitation by introducing a new ObjectIterator interface which
is common to all iterators: all it provides is access to the object ID
and name. It's thus sufficient to be used as input iterator for all of
the pipeline steps which do accept an input iterator.