blob: Skip CatfileInfo pipeline step in ListBlobs and ListAllBlobs
Depending on whether object data was requested in ListBlobs and
ListAllBlobs, we either only iterate over the `CatfileInfo()` pipeline
step to extrat information or we use `CatfileObject()` to also extract
the blobs' data. In the latter case we spawn `CatfileInfo()` regardless
of whether we actually require it though, and use it as intermediate
step between `Revlist()` and `CatfileObject()` that doesn't serve any
purpose.

Now that we can pass the `Revlist()` iterator into `CatfileObject()`
directly, we can get rid of this pipeline step altogether and thus
improve performance by getting rid of its process.

Changelog: performance