Use replica path when processing replication jobs
Praefect begins generating relative paths in 14.6. In order to ensure
replication RPCs end up targeting the correct repositories on the disk,
this commit changes the replicator to use the stored replica path for
the various RPCs. Renames and deletes are handled in the old manner as
they won't be produced anymore from 14.6 onwards and are left in place
only for backwards compatibility. Rest of the replication operations now
target the stored replica path instead of the relative path the clients
use to access the repository. As the replication jobs do not yet carry the
replica path in them, the field is filled from the database. A later commit
will begin including replica path in the replication jobs which will then
allow for removing GetReplicaPath.