Return the virtual storage parameter to GetPrimary
Praefect is now getting the primary of a repository by the repository
ID. Given that, the virtual storage paraemeter was no longer needed and
was removed. There are some tests for RepositoryReplicas in Rails' repo
which use the legacy election strategies. These legacy election strategies
still need the virtual storage parameter to determine the primary node.
In order to keep compatibility with the tests, this commit returns the
parameter to GetPrimary. While this works for now, we should really configure
Praefect properly in the tests to avoid being stuck with legacy implementations.