Support multiple outcomes in reconciliation tests
TestReconciler is currently flaky as some of the tests cases contain
randomness. In order to test these random cases, we need to be able
to assert against multiple sets of replication jobs. This allows for
asserting against the random outcome by defining the produced replication
jobs should match one of the expected sets.