sql: Provide timestamp values as argument to queries
We don't have a standard way of using timestamps in SQL
statements. We also use a time interval subtraction technic
that is pretty fragile. With that change we extract time
manipulation from SQL statements into Go code and pass
already prepared timestamps as parameters for the queries.
This also helps to solve problem of inability to get
different timestamps if SQL statements executed in single
transaction. It would be needed in the upcoming changes.