objectpool: Convert to use in-memory fetches
There is only a single callsite left in our codebase which creates
on-disk remotes, which is the `FetchIntoObjectPool` RPC. We do not need
this remote for anything though, and we have in the past made the move
to only ever use in-memory remotes.

Convert the implementation to use in-memory remotes.