objectpool: Convert to use fetches without remote
There is only a single callsite left in our codebase which creates
on-disk remotes, which is the `FetchIntoObjectPool` RPC. We do not need
this remote for anything though, and we have in the past made the move
to only ever use in-memory remotes.

Convert the implementation to fetch from the repository path directly
without creating a remote first. Note that this implicitly fixes a bug:
when creating a remote, Git sets up a default fetchspec which we use in
addition to the explicit fetchspec we pass to git-fetch(1), and as a
result we create two refs for each branch we fetch. Now that we don't
create a remote anymore this bug is fixed for good.

Changelog: fixed