sql: Reset sequences to initial value
We started to use sequences in expected results and
it is not always possible to predict a correct value
for the expected result. To get a determined results
we can restart sequence number generators to initial
values before test run. This change adds SequenceReset
method that does exactly it for all user defined
sequences in the database. It will be used in upcoming
changes.