sql: Do not truncate all tables
As we are gonna to pre-populate database with some data
and run tests on top of it we should not truncate full
database before each test run. The 'TruncateAll' method
calls were replaced with usage of transactions where it
is possible and a manual cleanup after test completion
where it is not possible to rely on the transactions. To
support transaction SQL type some helper functions were
refactored to accept 'Querier' interface instead of 'DB'
type and a 'Context' that is required by the interface.
In some cases the 'SequenceReset' method is called because
we use pre-defined values for the repository_id column and
need to reset sequence because they are not transactional.