sql: Truncate doesn't restart sequences
To get a purely new database schema for each test run
we used 'Truncate' method that cleans up tables and
resets sequences in the database to initial values.
As we are gonna to pre-populate database with some
data before test run we should not reset sequences
as it would result to duplicate keys during record
insertion operations. Those we drop sequence reset
and adjust the test with the change.