testhelper: Fix test configuration not printing errors on failure
When the testhelper fails to set up tests correctly, then it uses
`log.Fatalf("%v", err)` to abort the tests. This does not print the
error message though, where all you see is the unhelpful "exit code 1"
error. This is because we configure the log level to only print "panic"
messages, where `Fatalf()` will create a log message with "fatal" level.

Fix this by explicitly using `fmt.Printf()` to print the error and using
`os.Exit(1)`.