testhelper: Inline configuration of Ruby to break cyclic dependency
There are two callsites which configure the Ruby directory, once in the
gittest package and once in the testcfg package. Given that the helper
to do this is located in the testhelper package, it means that it needs
to depend on `internal/gitaly/config`.

Inline configuration of Ruby into both callers such that the mentioned
package may use the testhelper package. Note that this drops a call to
`cfg.ConfigureRuby()`: this call isn't required given that both
callsites already execute `cfg.Validate()`, which calls this function.

Note that this removes the ability to override the location of our Ruby
directory via the "GITALY_TEST_RUBY_DIR" environment variable. To the
best of my knowledge this variable is not used by anything, so removing
it is a nice cleanup while at it.