testhelper: Simplify configuration
Configuration of our test suite is still making use of a `sync.Once` due
to legacy reasons where `Configure()` was available to outside callers.
Now that it's a private function of the testhelper package it doesn't
make a whole lot of sense anymore because it's only invoked by
`testhelper.Run()`, which is only ever executed once.

Simplify the code by getting rid of the `sync.Once`.