testhelper: Don't clutter system temporary directory
During test execution, we create a lot of temporary files which we never
clean up, like for example the internal socket directory. While reigning
in on all of these "leaks" probably doesn't make a whole lot of sense,
it would still be nice to not leave behind so much garbage after test
runs. After a few hours of working, it's certainly impossible for me to
find anything of relevance in my "/tmp" directory because it's littered
with piles of Gitaly files.

Fix this by overriding the temporary directory during tests: instead of
creating them in the global temporary directory, we instead create them
in a "tmp" directory inside of our global test directory.