commit: Add test case exercising regression reading blob tree entries
We have recently converted the tree entries helpers to use the new
catfile interfaces. During this conversion, we have introduced a
regression when trying to read a tree entry which doesn't refer to a
subtree, but instead to a blob: we used to just ignore such tree entries
and return nothing, but now we return an error while trying to parse the
blob as a tree.

Add a testcase which exercises this code path to demonstrate the issue.