blob: Discard objects in case sending blob tree entries fails
We do not currently discard the objects read by `sendBlobTreeEntry()` in
case the function fails. As a result, the object reader would be dirty
after the function has failed and wouldn't be returned to the catfile
cache.

Discard the object's contents in case the function fails. While this
hasn't caused any known issues, it's still the right thing to do.