test: Check error returned by gRCP server
In the tests we start gRPC server to serve the
requests without verifying if it was started
successfully or not. If there was a problem we
should log it as it may contain a useful info
about the failure. The error won't be returned
in case the server is stopped because of the
Stop or Shutdown method called on it - the happy
scenario.