catfile: Fix deadlocking caused by early proces sdone notifications
Change the process-done condition to use locking, similar in spirit to
the preceding commit. This also fixes a race race condition observed in
CI where we are waiting forever for a catfile process to have been
returned to the cache because the notification triggered before we set
up the listener.

This commit also removes a useless deferred function call while at it.