commit: Fix test timeout compared by inefficient go-cmp diffing
Commit signature tests are frequently timing out in our CI, where a
single testcase frequently takes more than 400 seconds to succeed. Quite
surprisingly, the root cause is that we're using go-cmp to diff received
Proto messages. The diff algorithm used by go-cmp seems to be
inefficient enough to choke when comparing these messages, which contain
multiple megabytes of signed data.

Fix the issue by comparing the proto message's fields for equality
directly. On my machine, this decreases the runtime of this testcase
from more than 16 seconds to a fraction of a second. While this is not
the same numbers as in our CI, my machine is probably a lot more beefy.