datastore: Store replica path through CreateRepository
The `repositories` has a `replica_path` columnt that stores where the
replicas of the repository are stored on the disk. Currently the replicas
are always at the same path as `relative_path`. This will soon change as
Praefect will begin generating unique disk paths for the repositories. To
store these generated paths when a repository is created, this commit changes
CreateRepository to take in replica path as a parameter. The behavior still
stays the same and relative path is still set as the replica path.